requirements.txt
openai>=1.12.0
reportlab>=4.1.0
pydantic>=2.6.0
Pillow>=10.2.0
python-dotenv>=1.0.0
